Move time conversion func to utils
